win32: Trivially fix the compile error with printing.
_cairo_pattern_is_opaque() now takes the extents over which the
operation is defined so that we make exclude the clear pixels that
surround EXTEND_NONE surfaces when determining opacity. In order to take
full advantage of this we need to start performing an extents query on
the operation and pass that down to the analysis...

This patch however is just the quick compile fix to pass a NULL and
restore the old behaviour.

Fixes (with previous commit):
  Bug 26197 - Cairo doesn't build on windows
  http://bugs.freedesktop.org/show_bug.cgi?id=26197