[gl] Implement linear gradients acceleration.
This is significantly cribbed from Zach Laine's work, but reworked so
that gradients can be plugged in as either source or mask operands for
any of the paths.

This cuts the runtime of firefox-talos-svg in half on my GM45, at the
expense of gradient-alpha.  surface-pattern-operator also now fails
due to small rasterization differences.