gl: Add GLSL support for ARB_texture_rectangle, and repeat/reflect fallbacks.
Most testcases are now passing like the ARB_texture_non_power_of_two
case.  EXT_texture_rectangle support is dropped in favor of
ARB_texture_non_power_of_two.  If we have issues with drivers not
having that but having EXT (which just lacks the GLSL part of the
spec), we can split it out.  Right now non-GLSL support in cairo-gl is
probably in bad shape anyway and will require someone that cares for
it in order to get fixed up.