gl: Introduce _cairo_gl_device_has_glsl() function
The function takes a cairo_device_t argument as it can be called on an
unacquired device.

This is the first patch in a process to try to rid cairo-gl-surface.c of
casts from cairo_device_t to cairo_gl_context_t. As its members should
usually only be accessed when the device is locked, the only way the
cairo_gl_context_t should be acquired is using
_cairo_gl_context_acquire().