gl: Don't assert if there used to be a GL error
When acquiring the GL context, do not assert that the GL context is not
in an error state. Do not even call _cairo_error(). Handling GL errors
in other code is not Cairo's responsibility.

Instead just clear all previous errors so we don't accidentally set
surfaces into error states to unrelated errors.