XCB: Remove an incorrect clipping optimizations
It seems like the idea here was to optimize for the special case of a
rectangular clip, where the region operation could be replace by
direction operation on the draw extents. However, we cannot modify the
geometry for all operations, and in particular the glyphs. So remove
this from the common compositing path.

This fixes partial-clip-text in the test suite.

Signed-off-by: Uli Schlachter <psychon@znc.in>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>