XCB: Move the assert from 5a0f8f7320c916c
Calling _cairo_xcb_surface_ensure_picture() on a XCB surface whose fallback
member is non-null is always an error. It's possible that the surface first gets
a picture assigned and later it's fallback member is set. In this situation,
it's still wrong to use the surface's picture for any drawing-

Signed-off-by: Uli Schlachter <psychon@znc.in>